изменены pa у адресов упрощены методы автоизации/регистрации

+++ address/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 
 class Address(models.Model):
-    pa = models.PositiveIntegerField(verbose_name="Лицевой счет", primary_key=True)
+    pa = models.PositiveBigIntegerField(verbose_name="Лицевой счет", primary_key=True)
     street = models.CharField(verbose_name="Улица", max_length=64, blank=False)
     house = models.CharField(verbose_name="Дом", max_length=6, blank=True)
     building = models.CharField(verbose_name="Корпус", max_length=6, blank=True)
